Resolved CID 351725 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->m_spelltargetconstraints.size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Objects/ObjectMgr.cpp
@@ -3112,7 +3112,7 @@ void ObjectMgr::LoadSpellTargetConstraints()
 
     delete result;
 
-    LogNotice("ObjectMgr : Loaded constraints for %u spells...", m_spelltargetconstraints.size());
+    LogNotice("ObjectMgr : Loaded constraints for %u spells...", static_cast<uint32_t>(m_spelltargetconstraints.size()));
 }
 
 SpellTargetConstraint* ObjectMgr::GetSpellTargetConstraintForSpell(uint32 spellid)
